+++ package.json
@@ -14,7 +14,8 @@
     "firebase": "^12.7.0",
     "react": "^19.2.0",
     "react-dom": "^19.2.0",
-    "react-router-dom": "^7.11.0"
+    "react-router-dom": "^7.11.0",
+    "vite-plugin-pages": "^0.33.2"
   },
   "devDependencies": {
     "@eslint/js": "^9.39.1",
