+++ src/pages/template-registration.tsx
