+++ src/vite-env.d.ts
@@ -1,0 +1,7 @@
+/// <reference types="vite/client" />
+
+declare module '~react-pages' {
+  import type { RouteObject } from 'react-router-dom'
+  const routes: RouteObject[]
+  export default routes
+}
