+++ vite.config.ts
@@ -1,7 +1,13 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
+import Pages from 'vite-plugin-pages'
 
 // https://vite.dev/config/
 export default defineConfig({
-  plugins: [react()],
+  plugins: [
+    react(),
+    Pages({
+      dirs: 'src/pages',
+    }),
+  ],
 })
